// src/app/page/master/kotakabupaten/kotakabupaten.interface.ts
import { EventEmitter } from '@angular/core';
export interface KotaKabupaten {

  kdPropinsi?: any;
  kode?: any;
  kodeExternal?: string;
  namaExternal?: string;
  namaKotaKabupaten?: string;
  reportDisplay?: string;
  statusEnabled?: true;
  version?: any;
  noRec?: string;
  label?: any;
  value?: any;
  namaPropinsi?: string;

}